exercises1

+++ SUPER_EXTRA_4/Solution/ED_9.php
@@ -1,5 +1,5 @@
 <?php 
-    $num = readline();
+    $num = (int)readline();
     echo substr($num,0,1);
 
 ?>
